Move fetchSavedShippingAddresses into merlins connector
