+++ web/app/integration-manager/_merlins-connector/checkout/commands.js
@@ -2,7 +2,7 @@
 /* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
 /* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 
-import {makeJsonEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
+import {makeJsonEncodedRequest, makeRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
 import {SubmissionError} from 'redux-form'
 
 import {parseShippingInitialValues, parseLocations, parseShippingMethods, checkoutConfirmationParser} from './parsers'
@@ -219,4 +219,44 @@ export const submitPayment = (formValues) => (dispatch, getState) => {
                 throw new Error(responseJSON.message)
             }
         })
+}
+
+export const fetchSavedShippingAddresses = (isLoggedIn) => {
+    return (dispatch) => {
+        if (!isLoggedIn) {
+            return false
+        }
+
+        const fetchURL = `/rest/default/V1/carts/mine`
+        return makeRequest(fetchURL, {method: 'GET'})
+            .then((response) => response.json())
+            .then(({customer}) => {
+                let defaultShippingId
+                const addresses = customer.addresses.map((address) => {
+                    if (address.default_shipping) {
+                        defaultShippingId = address.id
+                    }
+
+                    // Not spreading `address` because it has key/values that
+                    // we want to rename and remove
+                    return {
+                        city: address.city,
+                        countryId: address.country_id,
+                        customerAddressId: `${address.id}`,
+                        customerId: `${address.customer_id}`,
+                        firstname: address.firstname,
+                        lastname: address.lastname,
+                        postcode: address.postcode,
+                        regionCode: address.region.region_code,
+                        regionId: `${address.region.region_id}`,
+                        region: address.region.region,
+                        street: address.street,
+                        telephone: address.telephone,
+                    }
+                })
+
+                dispatch(setDefaultShippingAddressId(defaultShippingId))
+                dispatch(receiveSavedShippingAddresses(addresses))
+            })
+    }
 }
