--- web/app/store/checkout/actions.js
@@ -1,8 +1,0 @@
-/* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
-/* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
-/* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
-
-import {createAction} from 'progressive-web-sdk/dist/utils/action-creation'
-
-export const receiveSavedShippingAddresses = createAction('Receive Saved Shipping Addresses', 'savedAddresses')
-
